phan: Add a plugin for handling "html template" files (#37585)
An "html template" file is a PHP file that produces a fragment of HTML,
loaded by a method or function call somewhere else in the codebase.

Phan will normally evaluate the template as being in the global scope,
meaning it doesn't know about `$this` or other variables and will
complain about attempts to access them.

This plugin adds support for `@html-template` and `@html-template-var`
annotations to declare that a file is a template, what its context is,
and what variables are intended to be available.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/9274442443

Upstream-Ref: Automattic/jetpack@1b8cb0ff93310bc82c8f2fbb62b000a3f2942827